uhhh... FOLKS.... let me tell you what we got going on up here
I have added an index.js file to my `db/` folder so that I can add associations between students and campuses; I've also written (AND TESTED👍🏻) the class methods so that they are pulling the associated campus/student with the requested student/campus. So I'll update my calls on the various components to get all the associated data as well, then update my JSX to render the relevant components of each.

I'll also need to get a Campus.getEnrollment method to see how many students are enrolled in that particular campus...

...and a BUNCH of other stuff I haven't even thought of yet
💥🧠💥