updated TWO things:
(1) the `single`___ views now pull the thing based on the ID, not on the index of the array.  That was a problem

(2) the seed file now has even numbers of things so as not to trick me like it did before by reversing the array and making the center thing look like the link was working correctly, even though it was inverted.